requirements.txt
Flask==1.1.2
h5py==2.10.0
idna==2.10
imageio==2.9.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.2.0
prompt-toolkit==3.0.18
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
tqdm==4.60.0

